Moved extensions from the Controls library/namespace to Microsoft.Toolkit.Uwp.UI.Extensions
