+++ Microsoft.Toolkit.Uwp.UI.Controls/Microsoft.Toolkit.Uwp.UI.Controls.csproj
@@ -135,12 +135,6 @@
     <Compile Include="OrbitView\OrbitViewPanel.cs" />
     <Compile Include="OrbitView\OrbitViewPanelItemArrangedArgs.cs" />
     <Compile Include="OrbitView\OrbitViewPanelItemsArrangedArgs.cs" />
-    <Compile Include="SurfaceDialTextboxHelper\SurfaceDialTextboxHelper.cs" />
-    <Compile Include="TextBoxMask\TextBoxMask.cs" />
-    <Compile Include="TextBoxMask\TextBoxMask.Properties.cs" />
-    <Compile Include="TextBoxRegex\TextBoxRegex.Data.cs" />
-    <Compile Include="TextBoxRegex\TextBoxRegex.Properties.cs" />
-    <Compile Include="TextBoxRegex\TextBoxRegex.cs" />
     <Compile Include="TileControl\TileControl.cs" />
     <Compile Include="WrapPanel\WrapPanel.cs" />
     <Compile Include="WrapPanel\WrapPanel.Data.cs" />
