+++ Microsoft.Toolkit.Uwp.UI/Extensions/TextBoxMask.Properties.cs
@@ -13,7 +13,7 @@
 using System.Collections.Generic;
 using Windows.UI.Xaml;
 
-namespace Microsoft.Toolkit.Uwp.UI.Controls
+namespace Microsoft.Toolkit.Uwp.UI.Extensions
 {
     /// <summary>
     /// TextBox mask property allows a user to more easily enter fixed width text in TextBox control
