+++ Microsoft.Toolkit.Uwp.UI/Extensions/TextBoxMask.cs
@@ -17,7 +17,7 @@ using Windows.ApplicationModel.DataTransfer;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 
-namespace Microsoft.Toolkit.Uwp.UI.Controls
+namespace Microsoft.Toolkit.Uwp.UI.Extensions
 {
     /// <summary>
     /// TextBox mask property allows a user to more easily enter fixed width text in TextBox control
